conflict solves and some modifications on KNN(not working yet)

+++ src/main/java/org/graphast/query/model/QueueEntry.java
@@ -2,7 +2,7 @@ package org.graphast.query.model;
 
 	public class QueueEntry implements Comparable<QueueEntry> {
 		private long id;
-		private short travelTime;
+		private int travelTime;
 
 		public QueueEntry(long id, int travelTime) {
 			this.id = id;
@@ -10,7 +10,7 @@ package org.graphast.query.model;
 		}
 
 		public int compareTo(QueueEntry another) {
-			return new Short(this.travelTime).compareTo(another.getTravelTime());
+			return new Integer(this.travelTime).compareTo(another.getTravelTime());
 		}
 		
 		@Override
@@ -30,7 +30,7 @@ package org.graphast.query.model;
 			this.id = id;
 		}
 
-		public short getTravelTime() {
+		public int getTravelTime() {
 			return travelTime;
 		}
